Protect against null project
git-svn-id: http://argouml.tigris.org/svn/argouml/trunk@16233 a161b567-7d1e-0410-9ef9-912c70fedb3f

+++ src/argouml-app/src/org/argouml/uml/diagram/ArgoDiagramImpl.java
@@ -595,8 +595,10 @@ public abstract class ArgoDiagramImpl extends Diagram
                 && "remove".equals(evt.getPropertyName())) {
 
             Model.getPump().removeModelEventListener(this, namespace, "remove");
-
-            getProject().moveToTrash(this);
+            
+            if (getProject() != null) {
+                getProject().moveToTrash(this);
+            }
         }
     }
